- Using the new functions. - Fixing security issues.

+++ src/Repositories/MessagesRepository.php
@@ -15,76 +15,85 @@ use Milebits\Society\Models\Message;
 class MessagesRepository extends ChildRepository
 {
     /**
+     * Create a new Message instance
+     *
      * @param Model $recipient
      * @param array $data
      * @return Model|Message|null
      */
     protected function newMessage(Model $recipient, array $data): ?Model
     {
-        return $this->all()->create(array_merge([
+        return $this->sent()->create(array_merge([
             'recipient_id' => $recipient->getKey(),
             'recipient_type' => $recipient->getMorphClass(),
         ], $data));
     }
 
     /**
+     * Send a message
+     *
      * @param Model $friend
+     * @param string|array|Collection|Arrayable|null $attachment
      * @param string $content
-     * @return Model|Message|null
+     * @return Model|Message
      */
-    public function sendTo(Model $friend, string $content)
+    public function send(Model $friend, string $content = '', $attachment = null): ?Message
     {
-        return $this->newMessage($friend, compact($content));
+        if (!$this->canSendMessageTo($friend)) return null;
+        $message = $this->newMessage($friend, compact($content));
+        if (!is_null($attachment)) {
+            if (is_string($attachment)) $attachment = [compact($attachment)];
+            if (is_array($attachment)) $attachment = collect($attachment);
+            if ($attachment instanceof Arrayable)
+                $message->attachments()->createMany($attachment->toArray());
+        }
+        return $message;
     }
 
     /**
-     * @param Model $friend
-     * @param string|array|Collection|Arrayable $attachment
-     * @param string $content
-     * @return Collection
-     */
-    public function sendAttachmentTo(Model $friend, $attachment, string $content = '')
-    {
-        if (is_string($attachment)) $attachment = [compact($attachment)];
-        if (is_array($attachment)) $attachment = collect($attachment);
-        return $this->sendTo($friend, $content)->attachments()->createMany($attachment->toArray());
-    }
-
-    /**
+     * Mark a message as delivered
+     *
      * @param Message $message
      * @return bool
      */
-    public function sendDeliveryNotificationFor(Message $message)
+    public function markAsDelivered(Message $message)
     {
         return $message->update(['delivered_at' => now()]);
     }
 
     /**
+     * Mark a message as seen
+     *
      * @param Message $message
      * @return bool
      */
-    public function sendSeenNotificationFor(Message $message)
+    public function markAsSeen(Message $message)
     {
         return $message->update(['seen_at' => now()]);
     }
 
     /**
+     * Delete a certain message
+     *
      * @param Message $message
      * @return bool|null
      * @throws Exception
      */
     public function delete(Message $message)
     {
+        if (!$this->canOperateMessage($message)) return false;
         return $message->delete();
     }
 
     /**
+     * Respond to a message
+     *
      * @param Message $message
      * @param string $content
      * @param string|null $attachment
      * @return Model|Message
      */
-    public function respondToMessage(Message $message, string $content, string $attachment = null)
+    public function respond(Message $message, string $content, string $attachment = null)
     {
         $message = $this->newMessage($message->sender()->first(), compact($content));
         $message->parentMessage()->associate($message);
@@ -94,6 +103,8 @@ class MessagesRepository extends ChildRepository
     }
 
     /**
+     * All messages sent or received by this user
+     *
      * @return Builder
      */
     public function all()
@@ -102,6 +113,8 @@ class MessagesRepository extends ChildRepository
     }
 
     /**
+     * Sent messages
+     *
      * @return MorphMany
      */
     public function sent()
@@ -110,10 +123,36 @@ class MessagesRepository extends ChildRepository
     }
 
     /**
+     * Received messages
+     *
      * @return MorphMany
      */
     public function received()
     {
         return $this->model()->morphMany(Message::class, 'recipient');
+    }
+
+    /**
+     * Check is whether the user can interact with this message and modify it's content
+     *
+     * @param Message $message
+     * @return bool
+     */
+    public function canOperateMessage(Message $message): bool
+    {
+        $sender = $message->sender()->first();
+        if (is_null($sender)) return false;
+        return $sender->is($this->model());
+    }
+
+    /**
+     * Check if whether this user can send a message to a user
+     *
+     * @param Model $person
+     * @return bool
+     */
+    public function canSendMessageTo(Model $person): bool
+    {
+        return !$this->society->friends->isBlockedBy($person);
     }
 }
